refactor(grpc): move Sync field_mask logic into the masking module
Keep the mask-interpretation logic out of the Sync service handlers. The
`BlockMask` strategy and its path parsing now live in `masking.rs` next to
`apply_mask`, with a module doc explaining when each applies: `apply_mask` is a
post-hoc JSON projection for low-volume Query responses, `BlockMask` is a
source-level projection for the high-volume Sync streams. The handlers only
extract the paths and consult the flags while building the version-specific
`AnyChainBlock`.
