+++ src/serve/grpc/masking.rs
@@ -1,3 +1,17 @@
+//! Field-mask strategies for gRPC responses.
+//!
+//! A `google.protobuf.FieldMask` lets a client request a subset of a response's
+//! fields. Two strategies for honoring it coexist here, picked per endpoint:
+//!
+//! - [`apply_mask`] is a *post-hoc projection*: it round-trips an already-built
+//!   response through JSON and keeps only the requested paths. It works for any
+//!   message but pays the cost of building (and serializing) fields the client
+//!   discards, so it suits low-volume Query responses.
+//! - [`BlockMask`] is a *source-level projection*: it interprets the mask up
+//!   front so the producer can skip building representations the client did not
+//!   request. This suits the high-volume Sync block streams, where parsing each
+//!   block into its `AnyChainBlock` representation is the dominant cost.
+
 use serde::{de::DeserializeOwned, Serialize};
 use serde_json::{json, Value};
 
@@ -44,11 +58,101 @@ where
     serde_json::from_value::<T>(new_json)
 }
 
+/// Source-level projection of an `AnyChainBlock` from a `FieldMask`.
+///
+/// `AnyChainBlock` carries two representations of the same block: the raw
+/// `native_bytes` and a fully parsed `chain` block. This mask records which of
+/// them a producer should populate, so the parsed representation is only built
+/// (and the raw bytes only copied) when actually requested.
+///
+/// Unlike [`apply_mask`], this does not touch a built response — the caller
+/// consults the flags while constructing the message.
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+pub struct BlockMask {
+    pub native_bytes: bool,
+    pub chain: bool,
+}
+
+impl BlockMask {
+    /// A mask that selects every representation.
+    pub const fn all() -> Self {
+        Self {
+            native_bytes: true,
+            chain: true,
+        }
+    }
+
+    /// Interprets a `FieldMask`'s paths against `AnyChainBlock`.
+    ///
+    /// An absent or empty mask (the caller passes `&[]`) selects everything, so
+    /// existing clients that send no mask are unaffected. Paths are matched
+    /// leniently: a leading `block.` segment (referring to the repeated `block`
+    /// field in the response message) is tolerated, as is a bare leaf name.
+    /// Recognized leaves are `native_bytes` and the parsed chain
+    /// (`cardano`/`chain`). Selecting the whole `block` field keeps both.
+    pub fn from_paths(paths: &[String]) -> Self {
+        if paths.is_empty() {
+            return Self::all();
+        }
+
+        let mut mask = Self {
+            native_bytes: false,
+            chain: false,
+        };
+
+        for path in paths {
+            // Tolerate the response-relative `block.` prefix.
+            let field = path.strip_prefix("block.").unwrap_or(path);
+            // Only the first segment matters for deciding what to populate.
+            match field.split('.').next().unwrap_or("") {
+                "native_bytes" => mask.native_bytes = true,
+                "cardano" | "chain" => mask.chain = true,
+                // The whole block (or an unrecognized empty path) keeps both.
+                "" | "block" => return Self::all(),
+                _ => {}
+            }
+        }
+
+        mask
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use serde::Deserialize;
 
     use super::*;
+
+    #[test]
+    fn block_mask_from_paths_semantics() {
+        // Absent/empty mask keeps everything.
+        assert_eq!(BlockMask::from_paths(&[]), BlockMask::all());
+
+        // Bare leaf names.
+        let bytes_only = BlockMask::from_paths(&["native_bytes".to_string()]);
+        assert!(bytes_only.native_bytes && !bytes_only.chain);
+
+        let chain_only = BlockMask::from_paths(&["cardano".to_string()]);
+        assert!(!chain_only.native_bytes && chain_only.chain);
+
+        // Response-relative `block.` prefix is tolerated.
+        let prefixed = BlockMask::from_paths(&["block.native_bytes".to_string()]);
+        assert!(prefixed.native_bytes && !prefixed.chain);
+
+        // Deeper paths into the parsed block select the chain representation.
+        let deep = BlockMask::from_paths(&["block.cardano.header".to_string()]);
+        assert!(!deep.native_bytes && deep.chain);
+
+        // Selecting the whole block keeps both.
+        assert_eq!(
+            BlockMask::from_paths(&["block".to_string()]),
+            BlockMask::all()
+        );
+
+        // Multiple paths accumulate.
+        let both = BlockMask::from_paths(&["native_bytes".to_string(), "cardano".to_string()]);
+        assert!(both.native_bytes && both.chain);
+    }
 
     #[derive(Deserialize, Serialize, Default, Debug, PartialEq, Clone)]
     #[serde(default)]
